requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.10.1
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
python-dotenv==1.0.1
email-validator==2.2.0
